feat: replace older canary component with standalone version
