+++ src/pages/canary.js
@@ -1,9 +1,9 @@
-import { CanaryStandalone } from "../components/CanaryStandalone";
+import { Canary } from "../components/Canary";
 
 export function CanaryPage({ url = "/canary/api", ...rest }) {
   return (
     <div className="border" {...rest}>
-      <CanaryStandalone url={url} topLayoutHeight={0} />
+      <Canary url={url} />
     </div>
   );
 }
